Use sentence object filler. Move simple_scrape to root dir and simplify parsers module so that files in it are easier to use from tests.
