--- utils/__init__.py
@@ -1,1 +1,0 @@
-from .fill_out_sentence_object import fill_out_sentence_object
